Created a new package called fedora.server.access.localservices which mirrors the Local & Built-in Services box on the system diagram
The class HttpService provides a way for mechanisms using the http protocol
to get their http content.

Method getHttpContent reads the contents of the specified
URL and returns the result as a MIMETypedStream.

This functionality was moved out of the FedoraAccessServlet and
into this new class to more closely model the system diagram.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@284 6cc838c6-e627-0410-8cce-a6db3432f928