moved collection of configuration info from constructor to static block since this info only needs to be obtained once from teh server.
Eventually, I think it makes sense to make DisseminationService a module and then this section of code would be moved into the init method f the module. I will make that change after beta1.

Need to do more in error handling for servlets but that will wait until after beta1. Problem is that servlet can't throw ServerException so errors in
the servlet need to be trapped some other way. Gets more complicated
for mechanisms which are doing external calls to DatastreamResolverServlet. When servlet can't resolve URL you can get
a variety of response codes like 404 or 500 which mechanism can choose
to handle or not to handle and then return something from the mechanism back to the originating call from Fedora so Fedora could get back 404 or 500
or even just an empty web page depending on how mechanism handles errors. This makes error dificault to trap on the Fedora server side.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@966 6cc838c6-e627-0410-8cce-a6db3432f928