changed the way database-related classes are cleaned up in finally block. Stress testing revealed some problems that did not manifest themselves at lower load levels.
I need to think about this section of code some more to be sure that this fix
doesn't leave any dangling database connnections, but it does seem to fix
the earlier problem I was seeing with loads of 5-10 concurrent users.

Essentially, the change leaves it up to the ConnectionPool to decide if and
when java.sql.Connection is closed. It does free the connection so it is once available to the ConnectionPool. It does close any ResultSets or Statements that have not already been garbage collected.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@1632 6cc838c6-e627-0410-8cce-a6db3432f928