added code to doCommit that now uses Chris' recent modifications to the search interface to catch deletions of dependent objects.
Code will now check if any other objects in the repository are dependent
(i.e., shared bDef or shared bMech) on the object about to be deleted.
If one or more related objects are found, then ObjectDependencyException
is thrown and the purge is aborted.

The user can then use the search inteface to obtain a list of objects that
depend on the one being deleted using the search criteria:

bDef~pid    - where pid is the PID of the object being deleted
and
bMech~pid


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@1645 6cc838c6-e627-0410-8cce-a6db3432f928