temporary fix for JDBC problem uncovered with MYSQL4 and version 3.x of mysql connector JDBC driver.
This works, but I want to take a closer look at the code to see if this situation occurs anyplace else and if there's a more general way of fixing the problem.

The problem has to do with nested ResultSets that are instantiated from the same Statement object.

According to the latest javadocs:

 By default, only one ResultSet object per Statement  object can be open at the same time. Therefore, if the reading of one ResultSet object is interleaved with the reading of another, each must have been generated by different Statement objects. All execution methods in the Statement interface implicitly close a statment's current ResultSet object if an open one exists.

However, implementation of the JDBC Driver may differ from how java.sql.Statement(Java 1.4.x) is defined which could explain why this did not cause problems with the earlier mysql connector and MYSQL 3.23.x

I think the safe way to handle this is to always have separate Statements when nesting ResultSets and then it should work regardless of the JDBC driver implementation.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@2077 6cc838c6-e627-0410-8cce-a6db3432f928