lockingUser=>ownerId
Got rid of the entire concept of "locking user" throughout the code
and replaced it with the concept of "object owner".  currently,
"object owner" defaults to fedoraAdmin (the only administrative user
supported by the software)... and the field ownerId replaces the
old lockingUser field in the object model, database, and search api.

Note: this change will require a fresh build and full ingest of objects,
as it changes a field's name in the database.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@2170 6cc838c6-e627-0410-8cce-a6db3432f928