PID generator is now backed by the database. This change ensures that if someone ingests an object with their own pid, and that pid is accepted due to a matching retainPids values, that the generator will never try to generate that PID.
This PID generator also automatically
upgrades the old one so that the previously
"highest generated pid#" is respected,
if present.

Example scenario, illustrating the necessity
of re-writing the pid generation stuff:

Say my repository is configured to generate
PIDs with the namespace "changeme",
and I also configure it to allow ingested
objects in the "changeme" namespace to
retain the pids embedded in them.

First, I ingest an object where the pid is
specified as empty.  The repository gives it
a PID of "changeme:1".  Next, I ingest another
object, only this time, I manually put the
PID, "changeme:2" in the METS.  Then I decide
to ingest another object, without a pre-specified
PID.  The PID generator previously would have
generated "changeme:2", and the commit would
have failed... because an object of that PID
already exists in the repository.

This is accomplished by having a method on
the PIDGenerator interface called
neverGeneratePID(String pid), which is called
anytime an ingest successfully bypasses the PID
generation process.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@2261 6cc838c6-e627-0410-8cce-a6db3432f928