improved error handling involving external datastreams that may contain broken URL references.
Previously, broken links would throw
"java.lang.reflect.InvocationTargetExceptionThe message was null" when ds.getContentStream() was called on the datastream when invoking default disseminator.

Instead of throwing an exception, this error will now generate a MIMETyped stream containing html explaining the exception and why the datastream could not be retrieved.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@2318 6cc838c6-e627-0410-8cce-a6db3432f928