Bug Fix   Bugzilla #13.
I believe there was some obsolete code involving the syntax of the WHERE
clause for Search queries where a special string " { escape '/'} " was appended
to the end of any WHERE clause string that contained special characters
that required escaping according to the SQL92 syntax rules. I think this was
used as a flag at some earlier point in the code to signal that the WHERE
clause string had characters that needed escaping prior to generating the SQL
command. However, the code in the toSQL method appears to properly
escape special characters like " '  ", " % ", etc... and also check for wildcard
characters like " * " and " ~ " so I think the special escape string is no longer
required. Currently, if there are special characters in the string, the characters
are properly escaped, but the special escape string is still left appended to the
end of the WHERE clause which thorws SQLException because the syntax
is invalid.

I have commented out the line of code that appends the escape string. If
we find that odd characters are not being properly handled in the search
string, we may need to revisit this problem, but in my initial testing it seems
to fix the problem of searching using the underscore character.

Reported error:
" Try the field search: cmodel~*E_D*
(this *should* return all docs with a SIMPLE_DOC content model)

The part where it generates the SQL query fails, but I don't know why.

If the backend db is mysql, it says something about the escape clause.

If the backend db is mckoi, it says

  Error querying sql db:
    Lexical error at line 1, column 71.
      Encountered: "{" (123), after : ""

Current recommendation is to not use underscores where possible, if they need
to be queried on. "


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@2571 6cc838c6-e627-0410-8cce-a6db3432f928