Bug Fix.  Bugzilla #9.
Changes the way that the OAI "Until" date parm is handled when using "Day"
granularity. If Day granularity is being used, the date value for the Until date parm
is modified to be Day + 23:59:59. This means that when using Day granualrity
specifying the saem Day for both from and until date parms will result in matching
all dates between 00:00:00 and 23:59:59 on that day.

Reported problem:

" For OAI requests, we're now interpreting both "from" and "until" parms
(which both have day-granularity) as meaning "midnight beginning that date".
This is correct for the "from" parm only.

We need instead to interpret a day-granularity "until" parm as meaning
"23:59:59pm -ending- that date".  That way, an OAI request with
equal day-granular "from" and "until" parms will return records for that date.

This accords with OAI spec and also with Chris's conversation with Naomi.
"


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@2588 6cc838c6-e627-0410-8cce-a6db3432f928