Bug Fix.  Bugzilla #24, #25.
Fixes a couple of bugs in replicator and did some general reformatting
and adding of comments to make updateComponents a little more readable.

#24 had to do with the improper updating of dsBind table when the binding
map of an existing object was altered. Result before fix was incomplete
update of dsBind which had the effect of breaking disseminations for that
object because of missing binding maps. The fix has worked on the test
cases I have tried, but should be banged on some more by others to be
sure there are no corner cases I overlooked in my test examples.

#25 had to do with the updating of dsBind table when creating or modifying
binding maps. Since dsBind table functions as both datastream and binding
table, its state field is that of the datastream and not of the binding itself.
Before fix, the state of the updated bindings was alwasy set to "A". I think
this was probably a leftover from earlier code where we could not change
state so state was always set to active. The fix now uses the state of the
datastream in the xml to set the state when creating or updating a binding
map. This guarantees that the state in dsBind table will always be the same
as that of datastream in xml.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@2604 6cc838c6-e627-0410-8cce-a6db3432f928