adds optional 7th argument to mysql-config script
if present, it will set character set and collation for mysql 4.1 databases
if omitted, same as it used to be so existing 3.23.x or 4.0.x users will notice no change

any value can be used for the 7th argument since script only checks to
see if 7th argument is present or not. Need to add to user docs.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@3197 6cc838c6-e627-0410-8cce-a6db3432f928