removed "subject" entry so matches AnySubject.
In the case of no-authN, cannot match on attributeId of backendService since that
attribute is unavailable. Only available when backendService authenticates.

After this change I am moving this policy out of 2.0 set of policies and
placing into other-example-repository-policies.

To mirror conditions of 2.0 out-of-the-box, there really should be
no policies in play for backend services and defaults in beSecurity.properties
would be basicAuth=false, ssl=false, and no IP restrictions.

Users wanting to secure backend services would need to use
the configuration tool to make existing backedn services secure.
Or we could provide a sample set of policies that do this for the
demo objects.....


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4047 6cc838c6-e627-0410-8cce-a6db3432f928