I modified getPort ad getProtocol to now dynamically return the correct port and protocol.
getPort is dependent on getProtocol and uses the value of
protocol to determine whether to get fedoraServerPort(unsecure)
r fedoraRedirectPort(secure)

getProtocol uses the settings in beSecurity to figure out if the
protocol is SSL or not. If SSL is in play, the fedoraInternalCall-1
entry should have callSSL=true.

It's a hack, but seems to return the correct values for SSL
depending on beSecurity.xml(beSecurity-defaults or beSecurity-secure)

I am now able to run the following API tests with success in secure mode:

TestAPIA - Ok
TestAPIM - Ok
TestAPIMLite -Ok
TestCommandLineUtilities - Ok
TestIngestDemoObjects - Ok

junit still failing for me just running TestAPIALite by itself.
Not sure if it is currently designed to be run independently or
whether it is dependent on configuration of TestAPIALiteOpen
and TestAPIALiteSecure....
I think Bill may still has some pending checkins...


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4362 6cc838c6-e627-0410-8cce-a6db3432f928