initial import
fedora-setup is designed to copy the appropriate config files into
place in the distribution. It takes a single argument which can have values
of:

secure-apim
secure-all
unsecure-apim
unsecure-all

I will be checking in mods to fedora-start shortly that checks if
the config files (fedora.fcfg, beSecurity, and web.xml) exist.
If not, it will abort startup and inform user they need to run fedora-setup
first.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4457 6cc838c6-e627-0410-8cce-a6db3432f928