This is one fix for the redirect problem with POSTs in Uplaoder.
The fix was to ping the uploadURL via HTTP GET first to see if
the URL is being redirected. If so, then pass the redirected location
onto the next POST request. I was hoping to replace most of the
HttpClient code with calls to FedoraClient methods but you can't
use FedoraClient.get(URL,boolean, boolean) here because this
method attempts to evaluate the redirected URL. This URL is
/management/upload which is handled by UploadServlet which
only accepts POSTS. For GETS it returns SC_BAD_REQUEST(400)
which causes the method to throw IOException because of the 400
error. Instead, I had to handle the HTTP GET call separately in
Uploader.

If anyone has a more elegant fix for this problem, jump in.....

This fixes the problem that was causing batchModify to fail
junit tests when attempting to upload a file.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4473 6cc838c6-e627-0410-8cce-a6db3432f928