removed code that runs fedora-setup in doSetup.
There are static variables in the parent classes that have paths to
fedora.fcfg so these variables are undefined even though
fedora-setup is run here in doSetup.

Safest way to assure that fedora.fcfg and other configs are in place
is to run fedora-setup prior to running any junit tests.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4513 6cc838c6-e627-0410-8cce-a6db3432f928