This checkin completes fix of bugzilla #147.
DefaultDOManager will now gracefully recover if the object
is somehow added to the registry but the insert fails for some
reason, throwing an exception but leaving the DB in a different
state.

Also, object, datastream, and disseminator labels are checked
for length when any related API-M method is called.

In addition, datastream IDs are validated when API-M:addDatastream
is called.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4888 6cc838c6-e627-0410-8cce-a6db3432f928