bugfix #182
This problem occurs when using the admin client bMech Builder and specifying
multiple mimetypes for a datastream in the DsInputSpec panel.

It is possible for the DSInputSpec to specify more than one mimetype for a
given datastream in the bMech Builder DSInputSpec panel by separating multiple
mimetpyes with a comma. However, upon serialization, the mimetypes are not
split out and the resulting serialization in the bMech object appears like:

<fbs:DSMIME>image/gif,image/jpeg</fbs:DSMIME>

This syntax can result in the mimetype not being recognized by the admin
client when attempting to bind datastreams because of the pattern mismatch.

Comma separated mimetypes should be serialized as separate elements in the
bmech object similar to:

<fbs:DSMIME>image/gif</fbs:DSMIME>
<fbs:DSMIME>image/jpeg</fbs:DSMIME>


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4984 6cc838c6-e627-0410-8cce-a6db3432f928