This adds the additional servlet mappings related to the AxisWSDLFilterServlet that was used as a fix along with Chris' new WSDL files for the RepoMMan/Rightscom folks.
I wanted to get this patch into cvs so it doesn't get lost in the shuffle. The eventual plan I think is to integrate the changes we implemented for the RepoMMan/Rightscom folks into the next release. This includes these changes to web.xml(now in cvs), addition of AxisWSDLFilterServlet(already in cvs), and addition of new refactored WSDL files(yet to be added to cvs).

These additions to web.xml should not affect other code and are only invoked when requesting the specific targets of access/wsdl or management/wsdl which are public URLs to access WSDL generated by Axis.

This can also serve as a reminder that we need to eventually integrate the new WSDL files that Chris created for RepoMMan patch and do final testing. So far there have been no reports of problems in using new wsdl and AxisWSDLServlet by repoMMan/Rightscom folks. We may want to wait until fedora war file changes are complete/stable before introducing the new wsdl files into cvs...


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4985 6cc838c6-e627-0410-8cce-a6db3432f928