bugfix #184
This bug only affects users who originally ingested objects in METS format and
then later exported the objects back out as METS. The bug prevented
reconstruction of the original METS dmdSecFedora/amdSec structure upon
export.

At ingest, this structural information for METS files is retained in the FORMAT_URI attribute when the files are converted to FOXML. Upon export
this info is then used to reconstruct the appropriate type of METS metadata
that was in the original file. If the object is born FOXML, there is no way to
distinguish metadata datastreams so when exported as METS, the code picks a
default METS metadata encoding (amdSec/techMD) and uses that encoding
for any X-type datastreams. The bug was that the reconstruction was never
being triggered upon export for METS objects so theresult was the same as
if the object was born FOXML as opposed to born METS.

The fix now allows the trigger to fire so born METS objects will retain their
original dmdSecFedora/amdSec encoding if later output as METS.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@4994 6cc838c6-e627-0410-8cce-a6db3432f928