Changed the contentDigest DIGEST and TYPE attributes from being required to optional.
This is to mirror how these attributes are treated in METS and to also allow the possibility of specifying TYPE only without specifying DIGEST. This is could be a common case if you wanted to specify the algorithm, but not the DIGEST and always allow the server to compute the DIGEST if autoChecksum is on.

As in METS, it does not make much sense to have the case where DIGEST is specified but TYPE is not, so I have included a schematron rule to enforce this condition.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@5197 6cc838c6-e627-0410-8cce-a6db3432f928