This is finishing up some of the policy work related to new checksum feature that Bob did not have a chance to finish before he left for Brazil.
Everything compiles, but I have not rigorously tested policies so this needs some additional testing. There are a couple questions regarding signature for compareDatastreamCehcksum that I will have to discuss with Bob when he is back after the holidays.

I wanted to go ahead and get this into cvs even though not rigorously tested so as not to hold up Sandy and her implementation of ownerID which will touch most of these classes...


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@5221 6cc838c6-e627-0410-8cce-a6db3432f928