fixes problem with setDatastreamVersionable policy
This policy attribute was different because it was the first attribute that was of type BooleanAttribute. All attributes until now had been dates or strings or string arrays.

Logic in fedora.server.MultiValueMap.set only checks for instanceofString, String[]s and null. For now, I changed DefaultAuthorization to pass the string equivalent of the boolean variable rather than change logic in MultiValueMap.

If we decide it is preferable to have DefaultAutorization to pass the boolean and then have MultiValueMap convert everything to strings, it should be an easy change.


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@5222 6cc838c6-e627-0410-8cce-a6db3432f928