recent changes in checksum related to normalizing inline-XML datastreams have changed the value of computed checksum for datastream DESC3 in this test object.
I have updated the checksum in the test object to match the value currently being generated so this object should now ingest successfully. With the previous value it was failing at ingest time because of checksum mismatch.

We may also want to consider moving ths test object out of the user demo object directory. Perhaps move it over to test directory with other test objects?


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@5393 6cc838c6-e627-0410-8cce-a6db3432f928